starting to move things into exporter lib, parameterize
